- Working coordinate find request - Some small fix

+++ app/src/main/java/com/example/myapplication/database/dao/CoordinateDAO.kt
@@ -21,7 +21,7 @@ interface CoordinateDAO : BaseDAO<Coordinate> {
     fun getById(coord_id: Long): Coordinate?
 
     @Query("select * from ${DatabaseConst.COORDINATE_TABLE_NAME} where longitude = :longitude and latitude = :latitude")
-    fun getByLongitudeAndLatitude(longitude: Int, latitude: Int): Coordinate?
+    fun getByLongitudeAndLatitude(longitude: Double, latitude: Double): Coordinate?
 
     @RawQuery
     fun getCoordsInRadius(query: SupportSQLiteQuery) : List<Coordinate>
